Integrate games helpers and globals into utils
