+++ src/components/Games/OptionMenu.jsx
@@ -1,7 +1,6 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import { Typography, IconButton, Tooltip, Menu, MenuItem } from '@material-ui/core';
-
 import MoreVertIcon from '@material-ui/icons/MoreVert';
 
 const propTypes = {
